explicit cast

+++ src/utils/src/Thread.c
@@ -69,7 +69,7 @@ PUBLIC_API STATUS defaultCreateThread(PTID pThreadId, startRoutine start, PVOID 
     STATUS retStatus = STATUS_SUCCESS;
 
 #if defined(KVS_DEFAULT_STACK_SIZE)
-    CHK_STATUS(defaultCreateThreadWithParams(pThreadId, start, KVS_DEFAULT_STACK_SIZE, args);
+    CHK_STATUS(defaultCreateThreadWithParams(pThreadId, start, (SIZE_T)KVS_DEFAULT_STACK_SIZE, args);
 #else
     CHK_STATUS(defaultCreateThreadWithParams(pThreadId, start, 0, args);
 #endif
@@ -218,9 +218,9 @@ PUBLIC_API STATUS defaultCreateThread(PTID pThreadId, startRoutine start, PVOID 
     STATUS retStatus = STATUS_SUCCESS;
 
 #if defined(KVS_DEFAULT_STACK_SIZE)
-    CHK_STATUS(defaultCreateThreadWithParams(pThreadId, start, KVS_DEFAULT_STACK_SIZE, args));
+    CHK_STATUS(defaultCreateThreadWithParams(pThreadId, start, (SIZE_T) KVS_DEFAULT_STACK_SIZE, args));
 #elif defined(CONSTRAINED_DEVICE)
-    CHK_STATUS(defaultCreateThreadWithParams(pThreadId, start, THREAD_STACK_SIZE_ON_CONSTRAINED_DEVICE, args));
+    CHK_STATUS(defaultCreateThreadWithParams(pThreadId, start, (SIZE_T) THREAD_STACK_SIZE_ON_CONSTRAINED_DEVICE, args));
 #else
     CHK_STATUS(defaultCreateThreadWithParams(pThreadId, start, 0, args);
 #endif
